This commit splits the tasks in app/signals/apps/signals/tasks.py into separate files, making it easier to manage and maintain the code.
