+++ app/signals/apps/signals/tasks/__init__.py
@@ -1,0 +1,24 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2023 Gemeente Amsterdam
+from signals.apps.signals.tasks.anonymize_reporter import anonymize_reporter, anonymize_reporters
+from signals.apps.signals.tasks.child_signals import (
+    apply_auto_create_children,
+    update_status_children_based_on_parent
+)
+from signals.apps.signals.tasks.clear_sessions import clearsessions
+from signals.apps.signals.tasks.delete_signals import delete_closed_signals
+from signals.apps.signals.tasks.refresh_database_view import (
+    refresh_materialized_view_public_signals_geography_feature_collection
+)
+from signals.apps.signals.tasks.signal_routing import apply_routing
+
+__all__ = [
+    'apply_auto_create_children',
+    'anonymize_reporter',
+    'anonymize_reporters',
+    'apply_routing',
+    'clearsessions',
+    'delete_closed_signals',
+    'refresh_materialized_view_public_signals_geography_feature_collection',
+    'update_status_children_based_on_parent',
+]
